# api/openapi.yaml
openapi: 3.1.0
info:
  title: Mesaar Logistics Operations API
  version: "1.0.0-draft"
  description: >
    Phase 2 contract draft. Reflects the EXISTING routes in `app/api/routes/*`
    remounted under the `/v1` prefix (ADR-005), plus clearly-tagged
    `driver-self (planned)` endpoints the delivered driver app needs
    (see docs/api-gap-analysis.md). Schemas mirror `app/schemas/*` and
    `app/models/*`. This file is hand-curated and is diffed in CI against the
    FastAPI-generated schema.
servers:
  - url: /v1
    description: Versioned base path
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate with email/password and receive a JWT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [auth]
      summary: Return the authenticated user's profile.
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserRead' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /drivers:
    post:
      tags: [drivers]
      summary: Create a driver profile (admin, manager).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DriverCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverRead' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      tags: [drivers]
      summary: List drivers (admin, manager).
      parameters:
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
      responses:
        '200':
          description: Page of drivers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DriverRead' }
  /drivers/{driver_id}:
    parameters:
      - { name: driver_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [drivers]
      summary: Get a driver by ID (admin, manager).
      responses:
        '200':
          description: Driver
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverRead' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [drivers]
      summary: Update a driver (admin, manager).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DriverUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverRead' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [drivers]
      summary: Delete a driver (admin, manager).
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /drivers/me:
    get:
      tags: [driver-self (planned)]
      summary: "[PLANNED] Profile of the authenticated driver."
      responses:
        '200':
          description: Driver profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverRead' }
    patch:
      tags: [driver-self (planned)]
      summary: "[PLANNED] Toggle own availability (online/offline)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { is_available: { type: boolean } }
              required: [is_available]
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverRead' }
...
  /shipments/{shipment_id}/accept:
    post:
      tags: [driver-self (planned)]
      summary: "[PLANNED] Driver self-accepts an offer (assigns self + default vehicle)."
      parameters:
        - { name: shipment_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Accepted/assigned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShipmentRead' }
        '400': { $ref: '#/components/responses/BadRequest' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Missing/invalid token
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Insufficient role
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    BadRequest:
      description: Validation/business-rule error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      properties:
        detail: { type: string }
      required: [detail]
    UserRole:
      type: string
      enum: [admin, manager, driver, client]
    ShipmentStatus:
      type: string
      enum: [created, ready, assigned, in_transit, delivered, cancelled, returned, failed]
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token: { type: string }
        token_type: { type: string, default: bearer }
        expires_in: { type: integer, description: seconds }
        user: { $ref: '#/components/schemas/UserRead' }
    UserRead:
      type: object
      required: [id, email, role, is_active, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string, nullable: true }
        role: { $ref: '#/components/schemas/UserRole' }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DriverRead:
      type: object
      required: [id, user_id, license_number, is_available, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        license_number: { type: string }
        license_class: { type: string, nullable: true }
        phone_number: { type: string, nullable: true }
        is_available: { type: boolean }
        home_warehouse_id: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DriverCreate:
      type: object
      required: [user_id, license_number]
      properties:
        user_id: { type: string, format: uuid }
        license_number: { type: string, minLength: 3, maxLength: 64 }
        license_class: { type: string, maxLength: 32, nullable: true }
        phone_number: { type: string, maxLength: 32, nullable: true }
        home_warehouse_id: { type: string, format: uuid, nullable: true }
    DriverUpdate:
      type: object
      properties:
        license_number: { type: string, minLength: 3, maxLength: 64 }
        license_class: { type: string, maxLength: 32, nullable: true }
        phone_number: { type: string, maxLength: 32, nullable: true }
        is_available: { type: boolean }
        home_warehouse_id: { type: string, format: uuid, nullable: true }
...
    ShipmentRead:
      type: object
      required: [id, reference_code, client_id, origin_warehouse_id, destination_warehouse_id, status, weight_kg, volume_m3, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        reference_code: { type: string }
        client_id: { type: string, format: uuid }
        origin_warehouse_id: { type: string, format: uuid }
        destination_warehouse_id: { type: string, format: uuid }
        driver_id: { type: string, format: uuid, nullable: true }
        vehicle_id: { type: string, format: uuid, nullable: true }
        status: { $ref: '#/components/schemas/ShipmentStatus' }
        weight_kg: { type: number }
        volume_m3: { type: number }
        pickup_at: { type: string, format: date-time, nullable: true }
        delivery_due_at: { type: string, format: date-time, nullable: true }
        delivered_at: { type: string, format: date-time, nullable: true }
        assigned_at: { type: string, format: date-time, nullable: true }
        cancelled_at: { type: string, format: date-time, nullable: true }
        failure_reason: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...